modifying rate datatype from float into integer

+++ database/migrations/2020_03_07_055626_create_cd_table.php
@@ -16,7 +16,7 @@ class CreateCdTable extends Migration
         Schema::create('cd', function (Blueprint $table) {
             $table->increments('id');
             $table->string('title');
-            $table->float('rate');
+            $table->integer('rate');
             $table->string('category');
             $table->integer('quantity');
         });
